ai-service/requirements.txt
# AttendAI — Python AI Service Requirements
# Install with: pip install -r requirements.txt
#
# NOTE: face_recognition requires cmake and dlib.
# On macOS: brew install cmake
# Then: pip install dlib (may take 5-10 minutes to compile)

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ── AI / Vision ────────────────────────────────────────────────────────────────
opencv-python==4.10.0.84
face-recognition==1.3.0
numpy==1.26.4

# ── Database ───────────────────────────────────────────────────────────────────
motor==3.5.1           # Async MongoDB driver (for asyncio)
pymongo==4.8.0         # Sync MongoDB driver

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1   # Load .env file
Pillow==10.4.0         # Image processing utilities
requests==2.32.3       # HTTP requests for downloading images from URLs
